+++ crates/star-cloudburst/src/lib/files/filedisplayinfo.rs
@@ -1,7 +1,8 @@
-use super::{FileTree, FileTreeDepthFirstIter, FileTreePathView, FlatFile, MetaV1FileRepr};
+use super::{FileTree, FileTreeDepthFirstIter, FlatFile, MetaV1FileRepr};
 use crate::metainfo::MetaV1;
 use std::{
-    iter::{self, Map, Once},
+    iter::{self, FusedIterator, Map, Once},
+    marker::PhantomData,
     num::NonZeroU64,
     slice::Iter,
 };
@@ -14,38 +15,24 @@ pub struct FileDisplayInfo<'file> {
     pub length: NonZeroU64,
 }
 
-pub enum FileDisplayInfoBranches<'iter> {
+/// Iterators that yield [FileDisplayInfo] based on the meta info dictionary version.
+pub(crate) enum FileDisplayInfoBranches<'iter> {
     /// Meta info version 1: single file
     MetaV1Once(Once<FileDisplayInfo<'iter>>),
     /// Meta info version 1: multiple files
+    #[allow(clippy::complexity)]
     MetaV1Multi(Map<Iter<'iter, FlatFile>, &'iter dyn Fn(&FlatFile) -> FileDisplayInfo>),
     /// Meta info version 2: single or multiple files
-    MetaV2(
-        Map<
-            FileTreeDepthFirstIter<'iter>,
-            &'iter dyn Fn(FileTreePathView) -> FileDisplayInfo<'iter>,
-        >,
-    ),
+    MetaV2(PathViewIntoDisplayInfoIter<'iter>),
 }
 
-fn flatfile_multi_to_filedisplay(flat_file: &FlatFile) -> FileDisplayInfo {
-    let mut file_path: Vec<_> = flat_file.path.iter().map(String::as_str).collect();
-    let name = file_path.remove(file_path.len() - 1);
-
-    FileDisplayInfo {
-        file_path,
-        name,
-        length: flat_file.length,
-    }
+/// Transform a collection of [FlatFile] or a [FileTree] into iterators of [FileDisplayInfo] without cloning owned values.
+pub(crate) trait AsFileDisplayInfo {
+    fn as_file_display(&self) -> FileDisplayInfoBranches<'_>;
 }
 
-/// Transform a collection of [FlatFile] or a [FileTree] into iterators of [FileDisplayInfo]
-pub trait IntoFileDisplayInfo {
-    fn into_file_display(&self) -> FileDisplayInfoBranches<'_>;
-}
-
-impl IntoFileDisplayInfo for MetaV1 {
-    fn into_file_display(&self) -> FileDisplayInfoBranches<'_> {
+impl AsFileDisplayInfo for MetaV1 {
+    fn as_file_display(&self) -> FileDisplayInfoBranches<'_> {
         match &self.files {
             &MetaV1FileRepr::Single(length) => {
                 FileDisplayInfoBranches::MetaV1Once(iter::once(FileDisplayInfo {
@@ -71,63 +58,58 @@ impl IntoFileDisplayInfo for MetaV1 {
     }
 }
 
-impl IntoFileDisplayInfo for FileTree {
-    // The anonymous lifetime is a subset of [FileTree]'s lifetime. [FileTreePathView]'s lifetime is a subset of
-    // the tree from which it borrows the path and file names. Therefore, [FileDisplayInfo]'s lifetime
-    // is a subset of that lifetime too since I'm just transferring ownership of the borrows.
-    fn into_file_display(&self) -> FileDisplayInfoBranches<'_> {
-        FileDisplayInfoBranches::MetaV2(self.iter_dfs().map(&|view| FileDisplayInfo {
+/// Iterator to map [FileTreePathView] => [FileDisplayInfo].
+///
+/// The iterator's lifetime is a subset of [FileTree]'s lifetime.
+/// [FileTreePathView]'s (which is yielded by the depth first iter) lifetime is a subset of
+/// the tree from which it borrows the path and file names. Therefore, [FileDisplayInfo]'s lifetime
+/// is a subset of that lifetime too since I'm just transferring ownership of the borrows.
+pub(crate) struct PathViewIntoDisplayInfoIter<'iter> {
+    phantom: PhantomData<FileTree>,
+    iter: FileTreeDepthFirstIter<'iter>,
+}
+
+impl<'iter> Iterator for PathViewIntoDisplayInfoIter<'iter> {
+    type Item = FileDisplayInfo<'iter>;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        let view = self.iter.next()?;
+        Some(FileDisplayInfo {
             file_path: view.directory.clone(),
             name: view.name,
             length: view.file_info.length,
-        }))
+        })
     }
 }
 
-pub struct FileDisplayInfoIter<'iter, I>
-where
-    I: Iterator<Item = FileDisplayInfo<'iter>>,
-{
-    iter: I,
+impl FusedIterator for PathViewIntoDisplayInfoIter<'_> {}
+
+impl AsFileDisplayInfo for FileTree {
+    #[inline]
+    fn as_file_display(&self) -> FileDisplayInfoBranches<'_> {
+        FileDisplayInfoBranches::MetaV2(PathViewIntoDisplayInfoIter {
+            phantom: PhantomData,
+            iter: self.iter_dfs(),
+        })
+    }
 }
 
-use crate::metainfo::MetaInfo;
-impl<'iter, I> FileDisplayInfoIter<'iter, I>
-where
-    I: Iterator<Item = FileDisplayInfo<'iter>>,
-{
-    fn file_display_info_iter(info: &MetaInfo) -> I {
-        match info {
-            MetaInfo::MetaV1(v1) => {
-                unimplemented!()
-            }
-            MetaInfo::MetaV2(v2) => {
-                unimplemented!()
-            }
-            MetaInfo::Hybrid(hybrid) => {
-                unimplemented!()
-            }
+pub struct FileDisplayInfoIter<'iter> {
+    pub(crate) branches: FileDisplayInfoBranches<'iter>,
+}
+
+impl<'iter> Iterator for FileDisplayInfoIter<'iter> {
+    type Item = FileDisplayInfo<'iter>;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        match &mut self.branches {
+            FileDisplayInfoBranches::MetaV1Once(iter) => iter.next(),
+            FileDisplayInfoBranches::MetaV1Multi(iter) => iter.next(),
+            FileDisplayInfoBranches::MetaV2(iter) => iter.next(),
         }
     }
 }
 
-/*impl IntoFileDisplayInfo for crate::metainfo::MetaV1 {
-    fn into_file_display<'file>(&'file self) -> FileDisplayInfo<'file> {
-        if let Some(files) = &self.files {
-            files.iter().map(|file| {
-                // let dir = std::iter::once(info.name.as_str()).chain(file.path.iter().map(|p|)
-
-                FileDisplayInfo { file_path: vec![] }
-            })
-        } else if let Some(file_length) = self.length {
-            FileDisplayInfo {
-                iter: std::iter::once(FileDisplayInfo {
-                    file_path: vec![],
-                    name: &info.name,
-                    length: file_length.get(),
-                }),
-            }
-        }
-    }
-}
-*/
+impl FusedIterator for FileDisplayInfoIter<'_> {}