+++ crates/star-cloudburst/src/lib/metainfo/metav2.rs
@@ -1,6 +1,6 @@
 use crate::{
     crypto::sha1::Sha1,
-    files::{FileDisplayInfo, FileTree},
+    files::FileTree,
     metainfo::serde_bool_int::{bool_from_int, bool_to_int},
     pieces::PieceLength,
 };